+++ lib/components/options/option_to_remove_expense_remittance/option_to_remove_expense_remittance_widget.dart
@@ -1,3 +1,4 @@
+import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
@@ -90,26 +91,42 @@ class _OptionToRemoveExpenseRemittanceWidgetState
                   hoverColor: Colors.transparent,
                   highlightColor: Colors.transparent,
                   onTap: () async {
-                    // remove from set
+                    if (valueOrDefault(currentUserDocument?.role, '') ==
+                        'admin') {
+                      // remove from set
 
-                    await widget.remittance!.update({
-                      ...mapToFirestore(
-                        {
-                          'transactions': FieldValue.arrayRemove(
-                              [widget.transaction?.reference]),
-                          'gross':
-                              FieldValue.increment(widget.transaction!.total),
-                          'net':
-                              FieldValue.increment(widget.transaction!.total),
-                          'expenses': FieldValue.increment(
-                              -(widget.transaction!.total)),
-                        },
-                      ),
-                    });
-                    // delete transaction
-                    await widget.transaction!.reference.delete();
-                    Navigator.pop(context);
-                    context.safePop();
+                      await widget.remittance!.update({
+                        ...mapToFirestore(
+                          {
+                            'transactions': FieldValue.arrayRemove(
+                                [widget.transaction?.reference]),
+                            'gross': FieldValue.increment(
+                                widget.transaction!.total),
+                            'net': FieldValue.increment(
+                                widget.transaction!.total),
+                            'expenses': FieldValue.increment(
+                                -(widget.transaction!.total)),
+                          },
+                        ),
+                      });
+                      // delete transaction
+                      await widget.transaction!.reference.delete();
+                      Navigator.pop(context);
+                      context.safePop();
+                    } else {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            'Consult Admin First!',
+                            style: TextStyle(
+                              color: FlutterFlowTheme.of(context).info,
+                            ),
+                          ),
+                          duration: const Duration(milliseconds: 4000),
+                          backgroundColor: FlutterFlowTheme.of(context).error,
+                        ),
+                      );
+                    }
                   },
                   child: Container(
                     width: double.infinity,
